requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
blinker==1.7.0
Brotli @ file:///private/tmp/python-brotli-20231117-5377-1dmlxkj/Brotli-1.1.0
build==1.0.3
cachelib==0.10.2
cachetools==5.3.2
certifi @ file:///private/tmp/python-certifi-20231117-4919-2f0o9m/certifi-2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
cs50==9.3.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
Flask==3.0.0
Flask-Dance==7.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
geographiclib==2.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.26.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.19.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.0
jupyter_core==5.6.1
jupyterlab_pygments==0.3.0
limits==3.7.0
mailtrap==2.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mutagen @ file:///private/tmp/python-mutagen-20231008-5554-1csby1/mutagen-1.47.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pip-tools==7.3.0
pipdeptree==2.13.1
platformdirs==4.1.0
pluggy==1.3.0
postgrest==0.13.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
redis==5.0.1
requests==2.32.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.3
termcolor==2.4.0
tinycss2==1.2.1
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
URLObject==2.4.3
wcwidth==0.2.12
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.16.0
zxcvbn==4.4.28
